[PRO-1892] add grpc endpoint for valuations and offers
